requirements.txt
python-dotenv==1.0.0
slack-bolt==1.18.0
requests==2.31.0
openai==0.28.1
